+++ docs/handbook/sources/build.sh
@@ -171,11 +171,13 @@ sed '1,19d' ../../../library/http/NOTES.md | pandoc -f gfm -t rst -o libraries/h
 sed '1,19d' ../../../library/http_client/NOTES.md | pandoc -f gfm -t rst -o libraries/http_client.rst
 sed '1,19d' ../../../library/http_cookies/NOTES.md | pandoc -f gfm -t rst -o libraries/http_cookies.rst
 sed '1,19d' ../../../library/http_cors/NOTES.md | pandoc -f gfm -t rst -o libraries/http_cors.rst
+sed '1,19d' ../../../library/http_directory_listing/NOTES.md | pandoc -f gfm -t rst -o libraries/http_directory_listing.rst
 sed '1,19d' ../../../library/http_multipart/NOTES.md | pandoc -f gfm -t rst -o libraries/http_multipart.rst
 sed '1,19d' ../../../library/http_router/NOTES.md | pandoc -f gfm -t rst -o libraries/http_router.rst
 sed '1,19d' ../../../library/http_server/NOTES.md | pandoc -f gfm -t rst -o libraries/http_server.rst
 sed '1,19d' ../../../library/http_session/NOTES.md | pandoc -f gfm -t rst -o libraries/http_session.rst
 sed '1,19d' ../../../library/http_socket/NOTES.md | pandoc -f gfm -t rst -o libraries/http_socket.rst
+sed '1,19d' ../../../library/http_static_files/NOTES.md | pandoc -f gfm -t rst -o libraries/http_static_files.rst
 sed '1,19d' ../../../library/http_websocket/NOTES.md | pandoc -f gfm -t rst -o libraries/http_websocket.rst
 sed '1,19d' ../../../library/http_websocket_messages/NOTES.md | pandoc -f gfm -t rst -o libraries/http_websocket_messages.rst
 sed '1,19d' ../../../library/http_websocket_session/NOTES.md | pandoc -f gfm -t rst -o libraries/http_websocket_session.rst