+++ paths/paths.pl
@@ -200,11 +200,13 @@ logtalk_library_path(http, library('http/')).
 logtalk_library_path(http_client, library('http_client/')).
 logtalk_library_path(http_cookies, library('http_cookies/')).
 logtalk_library_path(http_cors, library('http_cors/')).
+logtalk_library_path(http_directory_listing, library('http_directory_listing/')).
 logtalk_library_path(http_multipart, library('http_multipart/')).
 logtalk_library_path(http_router, library('http_router/')).
 logtalk_library_path(http_server, library('http_server/')).
 logtalk_library_path(http_session, library('http_session/')).
 logtalk_library_path(http_socket, library('http_socket/')).
+logtalk_library_path(http_static_files, library('http_static_files/')).
 logtalk_library_path(http_websocket, library('http_websocket/')).
 logtalk_library_path(http_websocket_messages, library('http_websocket_messages/')).
 logtalk_library_path(http_websocket_session, library('http_websocket_session/')).
@@ -408,6 +410,10 @@ logtalk_library_path(hanoi, threads('hanoi/')).
 logtalk_library_path(haunted_wasteland, examples('haunted_wasteland/')).
 logtalk_library_path(hello_world, examples('hello_world/')).
 logtalk_library_path(hooks, examples('hooks/')).
+logtalk_library_path(http_cookies_counter, examples('http_cookies_counter/')).
+logtalk_library_path(http_open_api, examples('http_open_api/')).
+logtalk_library_path(http_static_site, examples('http_static_site/')).
+logtalk_library_path(http_websocket_echo, examples('http_websocket_echo/')).
 logtalk_library_path(includes, examples('includes/')).
 logtalk_library_path(inheritance, examples('inheritance/')).
 logtalk_library_path(inlining, examples('inlining/')).