+++ paths/paths_libraries.pl
@@ -125,11 +125,13 @@ logtalk_library_path(http, library('http/')).
 logtalk_library_path(http_client, library('http_client/')).
 logtalk_library_path(http_cookies, library('http_cookies/')).
 logtalk_library_path(http_cors, library('http_cors/')).
+logtalk_library_path(http_directory_listing, library('http_directory_listing/')).
 logtalk_library_path(http_multipart, library('http_multipart/')).
 logtalk_library_path(http_router, library('http_router/')).
 logtalk_library_path(http_server, library('http_server/')).
 logtalk_library_path(http_session, library('http_session/')).
 logtalk_library_path(http_socket, library('http_socket/')).
+logtalk_library_path(http_static_files, library('http_static_files/')).
 logtalk_library_path(http_websocket, library('http_websocket/')).
 logtalk_library_path(http_websocket_messages, library('http_websocket_messages/')).
 logtalk_library_path(http_websocket_session, library('http_websocket_session/')).